+++ src/params.ts
@@ -2,11 +2,34 @@ import type { EmbedOptions, OpenOptions } from './interfaces';
 
 type Options = Omit<OpenOptions & EmbedOptions, 'origin' | 'newWindow' | 'height' | 'width'>;
 
+/**
+ * URL parameter names supported by the StackBlitz instance.
+ * Note that while updated instances perform a case-insensitive lookup
+ * for query parameters, some Enterprise Edition deployments may not,
+ * and we need to use specific (and sometimes inconsistent) casing;
+ * see for example 'hidedevtools' vs 'hideNavigation'.
+ */
+type ParamName =
+  | '_test'
+  | 'clicktoload'
+  | 'ctl'
+  | 'devtoolsheight'
+  | 'embed'
+  | 'file'
+  | 'hidedevtools'
+  | 'hideExplorer'
+  | 'hideNavigation'
+  | 'initialpath'
+  | 'showSidebar'
+  | 'terminalHeight'
+  | 'theme'
+  | 'view';
+
 const generators: Record<keyof Options, (value: any) => string> = {
   clickToLoad: (value: Options['clickToLoad']) => trueParam('ctl', value),
-  devToolsHeight: (value: Options['devToolsHeight']) => percentParam('devToolsHeight', value),
+  devToolsHeight: (value: Options['devToolsHeight']) => percentParam('devtoolsheight', value),
   forceEmbedLayout: (value: Options['forceEmbedLayout']) => trueParam('embed', value),
-  hideDevTools: (value: Options['hideDevTools']) => trueParam('hideDevTools', value),
+  hideDevTools: (value: Options['hideDevTools']) => trueParam('hidedevtools', value),
   hideExplorer: (value: Options['hideExplorer']) => trueParam('hideExplorer', value),
   hideNavigation: (value: Options['hideNavigation']) => trueParam('hideNavigation', value),
   showSidebar: (value: Options['showSidebar']) => booleanParam('showSidebar', value),
@@ -29,21 +52,21 @@ export function buildParams(options: Options = {}): string {
   return params.length ? `?${params.join('&')}` : '';
 }
 
-export function trueParam(name: string, value?: boolean): string {
+export function trueParam(name: ParamName, value?: boolean): string {
   if (value === true) {
     return `${name}=1`;
   }
   return '';
 }
 
-export function booleanParam(name: string, value?: boolean): string {
+export function booleanParam(name: ParamName, value?: boolean): string {
   if (typeof value === 'boolean') {
     return `${name}=${value ? '1' : '0'}`;
   }
   return '';
 }
 
-export function percentParam(name: string, value?: number): string {
+export function percentParam(name: ParamName, value?: number): string {
   if (typeof value === 'number' && !Number.isNaN(value)) {
     const clamped = Math.min(100, Math.max(0, value));
     return `${name}=${Math.round(clamped)}`;
@@ -51,14 +74,14 @@ export function percentParam(name: string, value?: number): string {
   return '';
 }
 
-export function enumParam(name: string, oneOf: string[], value?: string) {
+export function enumParam(name: ParamName, oneOf: string[], value?: string) {
   if (typeof value === 'string' && oneOf.includes(value)) {
     return `${name}=${value}`;
   }
   return '';
 }
 
-export function stringParams(name: string, value?: string | string[]): string[] {
+export function stringParams(name: ParamName, value?: string | string[]): string[] {
   const values = Array.isArray(value) ? value : [value];
   return values
     .filter((val) => typeof val === 'string' && val.trim() !== '')